refactor NcesParser and rake tasks
The import task now uses the gem: ar-extensions
to make importing much faster. This gem adds many 
useful extensions to AR. The feature used for importing
allows one statement to MySql to create many model
instances at once.

See:

  http://github.com/zdennis/ar-extensions/tree/master
  http://www.continuousthinking.com/tags/arext

This gem requirement has been added to the README
and to the rites Investigations codebase in the
with-portal branch.

The capability to import district and school data for 
just a selected list of States has been added and
if for example only data for RI and MA are imported
the entire import process now just takes 8m.

The rake tasks are now scoped to:

  rites_portal:setup:

instead of:

  rigse:setup:

There are new tasks:

  generate_nces_tables_migration
  generate_nces_indexes_migration

Just the table migration is checked in:

  20090701153613_create_nces06_tables.rb

Which is referenced in the RITES Investigations migration:

  20090701153613_create_nces06_tables.rb

Importing data can now be scoped to a list of states.

So when this task is run:

  import_nces_from_file

If the application where this engines is embedded
defines an array of state abbreviations and makes them 
available here:

  APP_CONFIG[:states_and_provinces]

example: ["RI", "MA"]

Then only districts and schools located in RI and MA
will be imported.

This constant is now defined:

  RITES_PORTAL_ROOT

So now when data files are downloaded by this task:

  download_nces_data

the files are saved to:

  RITES_PORTAL_ROOT/config/nces_data

The method that generates layouts has been refactored 
and split into one for districts and one for schools.

These two methods should either become just one method
again or the common code refactored out.

Using Ruby's String#unpack was about 10 times slower
than the original algorithm implemented in: