admins and managers can now switch user accounts
If you are logged in as a user with either admin
or manager roles you can now switch to any other
user -- except managers can't switch to user 
accounts with the admin role.

The "Switch" link appears in the to the right of the
'Preferences / Logout' links.

The view uses rails:

  option_groups_from_collection_for_select

helper to subdivide the select list of users available
to switch to. Users who you have switched to in this
session are at the top of the list under the heading
'recent'.