method: User#create_sds_counterpart won't create duplicates
The method:

  User#create_sds_counterpart

will no longer create duplicates.

This method can be run on all users without making duplicate sds
resources in order to add sds_config resources to legacy users.

It could probably be renamed to: find_or_create_sds_counterpart
but after legacy users are dealt with it will only 
be necessary as an after_create filter in User.