MIGRATION: User#default_user and site_admin
Add an attribute on the User model to indicate 
membership in the class of suspendable default users.

In addition add a site_admin attribute to make
generating views and performing actions for a site_admin
quicker in the future as we move towards permission_sets
based on membership models.

There is a new attribute required in the containing
applications config/settings.yml:

  enable_default_users

Normally this will be false for the production
environment.

The change in the portal engines environment.rb is
duplicated to the containing application.

There are three new class methods on the User model:

  User.default_users
  User.unsuspend_default_users
  User.suspend_default_users